feat: Implement foreground support
